requirements.txt
pandas>=1.5.0
requests>=2.31.0
python-dateutil>=2.8.2
pytz>=2023.3
duckdb==1.3.1
